3D side count algorithm implemented
Implemented algorithm, described in readme and previous commit, but for 3D spaces.
There are also needed little changes - I have written code for not only rectangle, but another voxel shapes too - but in task they are always in rectangle shape, so I need to rewrite code for this task. But at all code is fully functional.