Changed code for university project
In code I used standard syntax for reading lines and rows of matrix like for(string line : data), but project insists me to use first line of text file as row and lines count. For me it is not very useful, however it can be used if we want to analyze only certain part of figure.
Also all code was remade, so data firstly are separated to strings, which contain numbers, and after that are transferred to int via stoi(). Its better because now we dont need to iterate through data with i += 2 (+= 2 for space between numbers) and also it will allow to use non-one digit long number for height and grid parameters.